Fix defect: click same WBS twice resets timer
Previously, clicking the same WBS twice in a row reset the timer
as their was not time stopped and saved in between.

Now, a snapshot is createed whenever a WBS element is clicked
which can then be used to restart the timer. Additionally, the
state of the chronometer is stored to either start or stop the timer
depending on whether it was clicked the first or the second time