Dynamically add new WBS elements
New WBS elements can be added via a text + button
combination. The name is the text the user enters,
the ID will be generated.

The new WBS element is associated with the default
onClick listener that the others already use and thus
behaves the same.

Open: Design of the new elements as well as the list of WBS
elements